Update automatic_gec_tagging.py
# Changes made to provided better feedback during the exercise.

## U: type errors have a different text
Originally, if typing a fully compose sentence these "unnecessary"  error make sense, but in the slowly construction exercise where only 2 words might not be in the sentence the feedback doesn't quite reflect what's happening. To change this, I phrased it so it's a placement error, rather than being "unnecessary".

## Further changes were made to allow for better feedback
There were a few cases where the feedback would show old states as they were not cleaned up when re-checking the status. To fix this, whenever a token changes status the feedback property is re-writen.
Some cases the U: would be marked when not enough context was present. This is due to the comparison sentence only has +1 token than the original sentence. This is made to support iterative feedback, rather than just having a long list of Missing/Wrong matching tokens. To support this changes in the API were made so any tokens that are in the sentence have the property "isInSentence" which is True if the correct sentence has that token. In these cases, if the last token is marked with U: we do not set Feedback. (It might be correct, but something is missing in the sentence.)
- Instead of using "M:OTHER" the message will reflect the first unmerged label feedback. So if there were 2 things missing, it will point out the first component.

## Small changes to variable names
- Small typo in one of the variable names