Updates based on the feedback
Updated sklearn to version 1.4.0 (fixed some errors when utilizing pandas).
Refactored the utils code to remove repeated code. Added the .strip() as the final step in the remove_non_alphanumeric.
The stemmer is now loaded using the Languages in the database and checking if there is an NLTK Snowball stemmer for that language.
Renamed pre_process to stem_pre_process, to clearly state the intention of the method.
Fixed a typo "readibility" -> "readability".
Removed the "DifficultToRead" label.