Initializing the handler should only be done when retrieving the feed items
This seems to have caused a long slowdown of on the test suit. I imagine it was instanciatiating too many objects.

Runtime of tests went from 11 mins -> 2.5mins still a minute longer than the one without the feed. I believe this might have to do with "test_feed.py" changes which call the newspaper crawler.