Updates to tokenizer and update script
- The update script for Legacy bookmarks now does its best to find the context / bookmark pair in the article.
   - First it will try to find the context in the article.
   - Then it will try to find it in the title
   - If it does, then it will update the pointers for the text.

The bookmark can also result in a different tokenization, to address this, I try to match it in different ways within the context, if any of those succeed then I update the bookmark.

Currently, I log all the information of failed bookmarks for debugging purposes and to show situations where tokenization might have gone wrong.